requirements.txt
accelerate
aiofiles
aiohttp
aiosignal
altair
annotated-types
anyio
asgiref
asyncer
attrs
backoff
bcrypt
bidict
cachetools
certifi
chainlit
charset-normalizer
chroma-hnswlib
chromadb
click
colorama
coloredlogs
contourpy
ctransformers
cycler
dataclasses-json
Deprecated
fastapi
fastapi-socketio
ffmpy
filelock
filetype
flatbuffers
fonttools
frozenlist
fsspec
google-auth
googleapis-common-protos
gradio
gradio_client
greenlet
grpcio
h11
httpcore
httptools
httpx
huggingface-hub
humanfriendly
idna
importlib-metadata
importlib-resources
Jinja2
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
kiwisolver
kubernetes
langchain
langchain-community
langchain-core
langsmith
Lazify
markdown-it-py
MarkupSafe
marshmallow
matplotlib
mdurl
mmh3
monotonic
mpmath
multidict
mypy-extensions
nest-asyncio
networkx
nltk
numpy
oauthlib
onnxruntime
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
orjson
overrides
packaging
pandas
Pillow
posthog
protobuf
psutil
pulsar-client
py-cpuinfo
pyasn1
pyasn1-modules
pydantic
pydantic_core
pydub
Pygments
PyJWT
PyMuPDF
PyMuPDFb
pyparsing
pypdf
PyPDF2
PyPika
pyreadline3
python-dateutil
python-dotenv
python-engineio
python-graphql-client
python-multipart
python-socketio
pytz
PyYAML
referencing
regex
requests
requests-oauthlib
rich
rpds-py
rsa
safetensors
scikit-learn
scipy
semantic-version
sentence-transformers
sentencepiece
shellingham
simple-websocket
six
sniffio
SQLAlchemy
starlette
sympy
syncer
tenacity
threadpoolctl
tokenizers
tomli
tomlkit
toolz
torch
torchvision
tqdm
transformers
typer
typing-inspect
typing_extensions
tzdata
uptrace
urllib3
uvicorn
watchfiles
websocket-client
websockets
wrapt
wsproto
yarl
zipp
langfuse
langchain_openai
